-Page routes re-inserted
