+++ .babelrc.json
@@ -1,6 +1,6 @@
 {
   "presets": [
     "@babel/preset-env",
-    "@babel/preset-react"
+    ["@babel/preset-react", {"runtime":  "automatic"}]
   ]
 }
