--- src/components/BodyWrapper/BodyWrapper.js
@@ -1,9 +1,0 @@
-import "./BodyWrapper.scss"
-
-function BodyWrapper(props) {
-    return (
-        <div>{props.currentView}</div>
-    );
-}
-
-export default BodyWrapper;
