--- src/components/BodyWrapper/BodyWrapper.scss
@@ -1,11 +1,0 @@
-// Font-awesome
-@import url('https://use.fontawesome.com/releases/v5.0.7/css/all.css');
-
-// Functionality
-@import "src/stylesheets/variables";
-@import "src/stylesheets/mixins";
-html {
-  body {
-
- }
-}
