+++ src/pages/highscores/Highscores.scss
@@ -1,7 +1,96 @@
-$bgcolor: #333;
-$textcolor: #666;
+// Font-awesome
+@import url('https://use.fontawesome.com/releases/v5.0.7/css/all.css');
 
-.landing-body {
-	background-color: $bgcolor;
- 	color: $textcolor;
+// Functionality
+@import "src/stylesheets/variables";
+@import "src/stylesheets/mixins";
+
+.main-container-highscores {
+	.sub-container-highscores {
+
+		color: rgb(222, 222, 222);
+
+		h1 {
+			font-size: $font-xmed;
+			margin: 0rem;
+		}
+		h2 {
+			text-align: left;
+		}
+		nav {
+			padding: 1rem;
+			a {
+				padding: 1rem 3rem;
+				background-color: blue;
+				border-radius: 2px;
+
+				@include breakpoint-up(medium) {
+					padding: 1rem 4rem;
+				}
+			}
+			a:not(:last-child) {
+				margin-right: 1rem;
+			}
+		}
+		.skill-icon {
+			height: 7rem;
+		}
+		#hs-overall {
+			margin: 0rem 5rem 0rem 0rem;
+		}
+		.header-highscores {
+			text-align: left;
+			padding: 1rem 1rem 0rem 1rem;
+
+			@include breakpoint-up (large) {
+				padding: 0rem 1rem 0rem 4rem;
+			}
+		}
+		.sub-header-hs {
+			padding: 0rem 5rem;
+			margin-bottom: 1rem;
+		}
+		.filter-skill-hs {
+			padding: 3rem;
+		}
+		.filter-account-hs {
+			padding: .75rem;
+		}
+
+		#search-user-hs {
+			font-size: $font-sm;
+			padding: .75rem;
+			border: none;
+			border-radius: 2px;
+		}
+		table {
+			@include breakpoint-up (medium) {
+				width: 95vw;
+			}
+			@include breakpoint-up (large) {
+				width: 95vw;
+			}
+
+			th {
+				padding: .75rem;
+				background-color: green;
+			}
+			#rank {
+				text-align: left;
+				width: 5rem;
+			}
+			#player {
+				text-align: left;
+				width: 10rem;
+			}
+			#level, #exp {
+				text-align: right;
+				width: 10rem;
+			}
+			td {
+				padding: .5rem;
+				background-color: brown;
+			}
+		}
+	}
 }
